+++ src/App.js
@@ -1,6 +1,7 @@
 import React from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Favorites from "./js/Favorites";
+import Folder from "./js/Folder";
 
 function App() {
   return (
@@ -8,6 +9,7 @@ function App() {
       <BrowserRouter>
         <Routes>
           <Route path="/favorites" element={<Favorites />}></Route>
+          <Route path="/folder" element={<Folder />}></Route>
         </Routes>
       </BrowserRouter>
     </div>
