+++ src/js/Favorites.js
@@ -1,28 +1,11 @@
 import React, { useState, useEffect } from "react";
 import "../css/Favorites.css";
+import Nav from "./Nav";
 
 function Favorites() {
-  const [profile, setProfile] = useState();
   const [folder, setFolder] = useState(null);
 
   useEffect(() => {
-    const fetchUserData = async () => {
-      try {
-        const response = await fetch(
-          "https://bootcamp-api.codeit.kr/api/sample/user"
-        );
-        const data = await response.json();
-        if (data) {
-          setProfile(data);
-        } else {
-          setProfile();
-        }
-      } catch (error) {
-        console.error("User Error:", error);
-        setProfile();
-      }
-    };
-
     const fetchFolderData = async () => {
       try {
         const response = await fetch(
@@ -40,7 +23,6 @@ function Favorites() {
       }
     };
 
-    fetchUserData();
     fetchFolderData();
   }, []);
 
@@ -82,29 +64,7 @@ function Favorites() {
 
   return (
     <>
-      <nav>
-        <div className="gnb">
-          <a href="index.html">
-            <img src="./images/logo.svg" alt="홈으로 연결된 Linkbrary 로고" />
-          </a>
-          {profile ? (
-            <div className="FavoritesProfileContent">
-              <img
-                src={profile.profileImageSource}
-                alt="프로필 이미지"
-                className="FavoritesProfileImgSrc"
-              />
-              <a className="FavoritesProfileEmail" href="profile.html">
-                <span>{profile.email}</span>
-              </a>
-            </div>
-          ) : (
-            <a className="cta cta-short" href="signin.html">
-              <span>로그인</span>
-            </a>
-          )}
-        </div>
-      </nav>
+      <Nav />
       {folder && (
         <div className="FavoritesProfile">
           <img
