+++ src/js/Folder.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Folder() {
+  <></>;
+}
+
+export default Folder;
